requirements.txt
streamlit
python-dotenv
google-generativeai
requests
Pillow
duckduckgo-search
